requirements.txt
tweepy==4.14.0
deep-translator==1.11.4
python-dotenv==1.0.1
requests==2.32.3
